requirements.txt
SQLAlchemy==2.0.17
python-telegram-bot[all]==20.3
asyncpg==0.27.0
gspread==5.9.0
oauth2client==4.1.3